+++ crates/try_v2_derive/tests/compilation/examples/pass_MultipleFields.rs
@@ -1,0 +1,15 @@
+#![feature(never_type)]
+#![feature(try_trait_v2)]
+#![feature(try_trait_v2_residual)]
+
+use try_v2_derive::{Try, Try_ConvertResult};
+
+#[derive(Debug, Try, Try_ConvertResult)]
+#[allow(unused)]
+#[must_use]
+enum MultipleFields<T> {
+    Ok(T),
+    OtherError(String, String, i32),
+}
+
+fn main() {}