+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog try_v2
 
+## [v0.5.0]
+
+### New features
+
+- Add `trait Transform` (Forces move of derive macros to own crate & re-export)
+
+### Breaking changes
+
+- Removed derive `Try_Methods` in favour of trait-based implementation
+
 ## [v0.4.2]
 
 ### Bugfixes